requirements.txt
# requirements.txt
streamlit
ultralytics
torch
torchvision
pillow
opencv-python
numpy